122: add contract id to project
Most projects that you are recoding the time for will be billed in some form.
Basis of such a project is a contract, which is going to provide some sort
of foreign billing id or contract id that has to be given when reporting
to the contractee.

The contract id:
* is part of json exports
* is part of reports
* is a property of the project

Resolved #122.